requirements.txt
python-chess=0.23.10
numpy=1.15.4
